requirements.txt
aiosmtplib==1.1.7
amqp==5.1.1
anyio==3.6.2
arabic-reshaper==2.1.3
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.0.5
async-timeout==4.0.2
autopep8==1.6.0
awscli==1.27.35
awsebcli==3.20.3
backcall==0.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
botocore==1.23.54
bs4==0.0.1
celery==5.2.7
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
channels==4.0.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
cryptography==37.0.2
cssselect2==0.6.0
debugpy==1.6.0
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.2
django-admin-honeypot==1.1.0
django-admin-thumbnails==0.2.6
django-allauth==0.51.0
django-bootstrap4==3.0.1
django-celery-beat==2.4.0
django-celery-results==2.4.0
django-ckeditor==6.5.1
django-cleanup==6.0.0
django-crispy-forms==1.9.2
django-crontab==0.7.1
django-filter==22.1
django-heroku==0.3.1
django-import-export==2.7.1
django-ipware==4.0.2
django-js-asset==2.0.0
django-multiselectfield==0.1.12
django-phonenumber-field==6.1.0
django-session-timeout==0.1.0
django-timezone-field==5.0
django-utils-six==2.0
django-widget-tweaks==1.4.8
djangorestframework==3.13.1
docutils==0.16
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
future==0.16.0
gunicorn==20.1.0
h11==0.12.0
heroku==0.1.4
html5lib==1.1
httpcore==0.13.7
httpx==0.19.0
idna==3.3
iexfinance==0.5.0
ipykernel==6.14.0
ipython==8.4.0
jedi==0.18.1
jmespath==0.10.0
jupyter-client==7.3.4
jupyter-core==4.10.0
kombu==5.2.4
lxml==4.9.0
Markdown==3.4.1
MarkupPy==1.14
matplotlib-inline==0.1.3
messages==0.8.0
MouseInfo==0.1.3
mysqlclient==2.1.1
nest-asyncio==1.5.5
numpy==1.23.4
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
oscrypto==1.3.0
packaging==21.3
pafy==0.5.5
pandas==1.5.1
parso==0.8.3
pathspec==0.9.0
phonenumbers==8.12.48
pickleshare==0.7.5
Pillow==9.0.0
prompt-toolkit==3.0.29
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pure-eval==0.2.2
pyaes==1.6.1
pyasn1==0.4.8
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.15.0
PyGetWindow==0.0.9
Pygments==2.12.0
pyHanko==0.13.1
pyhanko-certvalidator==0.19.5
PyJWT==1.7.1
PyMsgBox==1.0.9
pyparsing==3.0.9
PyPDF3==1.0.6
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
Pyrogram==2.0.26
PyScreeze==0.1.28
PySocks==1.7.1
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==1.5
python-decouple==3.6
python-dotenv==0.19.2
python3-openid==3.2.0
pytube==12.1.0
pytube3==9.6.4
pytweening==1.0.4
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
pywin32==304
PyYAML==5.4.1
pyzmq==23.1.0
qrcode==7.3.1
qrtools==0.0.2
razorpay==1.2.0
redis==4.3.4
reportlab==3.6.10
requests==2.26.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.7.2
s3transfer==0.6.0
semantic-version==2.8.5
simplegeneric==0.8.1
six==1.14.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
stripe==3.5.0
svglib==1.3.0
tablib==3.2.0
termcolor==1.1.0
TgCrypto==1.2.3
tinycss2==1.1.1
tmdbsimple==2.9.1
toml==0.10.2
tornado==6.1
tqdm==4.64.0
traitlets==5.2.2.post1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.13
validus==0.3.0
vine==5.0.0
waitress==2.1.2
wcwidth==0.1.9
webencodings==0.5.1
websockets==10.3
whitenoise==5.1.0
wrapt==1.14.1
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
yapf==0.32.0
youtube-dl==2020.12.14
